requirements.txt
Flask
google-generativeai # Replace with the actual version if needed
flask_limiter
